// client/src/api-client/models/CreateTrip.ts
/* tslint:disable */
/* eslint-disable */
/**
 * journly-server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateTrip
 */
export interface CreateTrip {
    /**
     * 
     * @type {Date}
     * @memberof CreateTrip
     */
    endDate?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof CreateTrip
     */
    startDate?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof CreateTrip
     */
    title?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateTrip
     */
    userId: string;
}

/**
 * Check if a given object implements the CreateTrip interface.
 */
export function instanceOfCreateTrip(value: object): value is CreateTrip {
    if (!('userId' in value) || value['userId'] === undefined) return false;
    return true;
}

export function CreateTripFromJSON(json: any): CreateTrip {
    return CreateTripFromJSONTyped(json, false);
}

export function CreateTripFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateTrip {
    if (json == null) {
        return json;
    }
    return {
        
        'endDate': json['end_date'] == null ? undefined : (new Date(json['end_date'])),
        'startDate': json['start_date'] == null ? undefined : (new Date(json['start_date'])),
        'title': json['title'] == null ? undefined : json['title'],
        'userId': json['user_id'],
    };
}

export function CreateTripToJSON(json: any): CreateTrip {
    return CreateTripToJSONTyped(json, false);
}

export function CreateTripToJSONTyped(value?: CreateTrip | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'end_date': value['endDate'] == null ? undefined : ((value['endDate'] as any).toISOString().substring(0,10)),
        'start_date': value['startDate'] == null ? undefined : ((value['startDate'] as any).toISOString().substring(0,10)),
        'title': value['title'],
        'user_id': value['userId'],
    };
}